+++ app/src/main/java/br/ufpb/dcx/appalpha/control/service/interfaces/ThemesApiServiceInterface.java
@@ -9,6 +9,6 @@ import retrofit2.http.Path;
 public interface ThemesApiServiceInterface {
 
     @GET("/educapi/v1/api/contexts/{idContext}")
-    Call<Theme> find(@Path("idContext") int idContext);
+    Call<Theme> find(@Path("idContext") long idContext);
 
 }
